# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "BCGPT - Complete 3-Layer Platform API",
    "description": "Complete platform with 3 layers: BCGPT (Data - Basecamp integration), Flow (Execution - Automation via Activepieces), PMOS (Intelligence - Predictions, patterns, insights). 30+ essential operations in OpenAPI, 300+ tools via mcp_call. See x-assistant-instructions for tool-usage policy.",
    "version": "2.0.0"
  },
...
  "servers": [
    {
      "url": "https://bcgpt.wickedlab.io",
      "description": "Production server"
    }
  ],
  "components": {
...
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key from https://bcgpt.wickedlab.io/connect — paste this key when ChatGPT asks for credentials."
      }
    }
  },
  "security": [],
  "paths": {
...
    "/action/pmos_predict_completion": {
      "post": {
        "summary": "[PMOS] Predict project completion date based on velocity and remaining work",
        "operationId": "pmos_predict_completion",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "api_key": {
                    "type": "string",
                    "description": "Your personal API key"
                  },
                  "project_id": {
                    "type": "string",
                    "description": "Basecamp project ID"
                  }
                },
                "required": [
                  "project_id"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/action/pmos_insights_list": {
      "post": {
        "summary": "[PMOS] Get actionable insights generated by intelligence layer",
        "operationId": "pmos_insights_list",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "api_key": {
                    "type": "string",
                    "description": "Your personal API key"
                  },
                  "limit": {
                    "type": "integer",
                    "description": "Maximum insights to return",
                    "default": 50
                  },
                  "acknowledged": {
                    "type": "boolean",
                    "description": "Show acknowledged insights",
                    "default": false
                  }
                },
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
    "/action/flow_list": {
      "post": {
        "summary": "[Flow] List all automation flows",
        "operationId": "flow_list",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "api_key": {
                    "type": "string",
                    "description": "Your personal API key"
                  },
                  "project_id": {
                    "type": "string",
                    "description": "Optional Activepieces project ID filter"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/action/flow_trigger": {
      "post": {
        "summary": "[Flow] Manually trigger a flow with optional payload",
        "operationId": "flow_trigger",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "api_key": {
                    "type": "string",
                    "description": "Your personal API key"
                  },
                  "flow_id": {
                    "type": "string",
                    "description": "Flow ID to trigger"
                  },
                  "payload": {
                    "type": "object",
                    "description": "Data to pass to the flow"
                  }
                },
                "required": [
                  "flow_id"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
  }
}